avoid installation of golang when there already exists.

+++ setup.py
@@ -63,8 +63,8 @@ def prepare_rust():
     #     exec_shell_cmd("echo '%s' >> ~/.bashrc" % rustPath)
 
 def prepare_golang():
-    exit(-1)
-    exec_shell_cmd("sudo apt-get install -y golang")
+    if not os.path.exists("/usr/bin/go"):
+        exec_shell_cmd("sudo apt-get install -y golang")
 
 def prepare_shadow_dependencies():
     exec_shell_cmd("sudo apt-get install libcurl4-openssl-dev")
